+++ dev/com.ibm.ws.jsf.2.2_fat/test-applications/JSF22FacesFlows.war/src/com/ibm/ws/jsf22/fat/flows/beans/jsf22/InitializerBean.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  */
-package com.ibm.ws.jsf22.fat.flows.beans;
+package com.ibm.ws.jsf22.fat.flows.beans.jsf22;
 
 import java.util.Map;
